PSR-4 autoloading compliance + other fixes/tweaks/updates
Needs more testing.
