+++ examples/mandates.php
@@ -1,10 +1,10 @@
-<?php 
+<?php
 
 require_once __DIR__.'/iniatialization.php';
 
 /* Testing mandates */
-/** 
- * Creating a Mandate 
+/**
+ * Creating a Mandate
  * */
 
 
@@ -12,7 +12,7 @@ $order_id = "1234";
 $customer_id = "5678";
 
 $mandate_id = $this->bluem->CreateMandateId(
-	$order_id, 
+	$order_id,
 	$customer_id
 );
 
@@ -22,11 +22,11 @@ $response = $this->bluem->Mandate(
 	$mandate_id
 );
 
-if (is_a($response, "Bluem\BluemPHP\ErrorBluemResponse", false)) 
+if (is_a($response, "Bluem\BluemPHP\Responses\ErrorBluemResponse", false))
 {
 	throw new Exception(
-		"An error occured in the payment method. 
-		Please contact the webshop owner with this message:  " . 
+		"An error occured in the payment method.
+		Please contact the webshop owner with this message:  " .
 		$response->error()
 	);
 }
@@ -35,7 +35,7 @@ $attrs = $response->EMandateTransactionResponse->attributes();
 
 if (!isset($attrs['entranceCode'])) {
 	throw new Exception(
-		"An error occured in reading the transaction response. 
+		"An error occured in reading the transaction response.
 		Please contact the webshop owner"
 	);
 }
@@ -47,14 +47,14 @@ $entranceCode = $attrs['entranceCode'] . "";
 
 if (isset($response->EMandateTransactionResponse->TransactionURL)) {
 	$transactionURL = ($response->EMandateTransactionResponse->TransactionURL . "");
-	// @todo redirect to the above transaction URL and save the initiated transaction 
+	// @todo redirect to the above transaction URL and save the initiated transaction
 
-} else { 
+} else {
 	// @todo no proper status given, show an error.
 }
 
 
-/** 
+/**
  * Requesting a Mandate status
  * */
 
@@ -76,16 +76,16 @@ $statusCode = $statusUpdateObject->EMandateStatus->Status . "";
 if ($statusCode === "Success") {
 
 	// @todo do what you need to do to mark the transaction completed on your end
-	
+
 } elseif ($statusCode === "Cancelled") {
-	
+
 	// @todo do what you need to when the transaction has been cancelled
 
 } elseif ($statusCode === "Open" || $statusCode == "Pending") {
-	
+
 	// the transaction is still open. it might take some time to process
 	// @todo show a message that reflects this
-	
+
 } elseif ($statusCode === "Expired") {
 
 	// @todo show a message that reflects an expired transaction;
