+++ src/Requests/EmandateBluemRequest.php
@@ -1,71 +1,9 @@
 <?php
 
-/*
- * (c) 2020 - Daan Rijpkema <info@daanrijpkema.com>
- *
- * This source file is subject to the license that is bundled
- * with this source code in the file LICENSE.
- */
+namespace Bluem\BluemPHP\Requests;
 
-namespace Bluem\BluemPHP;
-
+use Bluem\BluemPHP\Contexts\MandatesContext;
 use Carbon\Carbon;
-
-/**
- * 	EMandateStatusRequest
- */
-class EmandateStatusBluemRequest extends BluemRequest
-{
-    public $typeIdentifier = "requestStatus";
-    public $request_url_type = "mr";
-    public $transaction_code = "SRX";
-
-
-    public function __construct(
-        $config, $mandateID, $entranceCode="", $expected_return=""
-    ) {
-        parent::__construct(
-            $config, $entranceCode, $expected_return
-        );
-
-        $this->xmlInterfaceName = "EMandateInterface";
-        $this->typeIdentifier = "requestStatus";
-
-        $this->mandateID = $mandateID;
-
-        $this->context = new MandatesContext(
-            $config->localInstrumentCode
-        );
-    }
-
-    public function TransactionType()
-    {
-        return "SRX";
-    }
-
-    public function XmlString() : String
-    {
-        return $this->XmlRequestInterfaceWrap(
-            $this->xmlInterfaceName,
-            'StatusRequest',
-            $this->XmlRequestObjectWrap(
-                'EMandateStatusRequest',
-                '<MandateID>'.$this->mandateID.'</MandateID>'
-            )
-        );
-
-        /* // Reference
-                return '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-        <EMandateInterface xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" type="StatusRequest"
-            mode="direct" senderID="'.$this->senderID.'" version="1.0" createDateTime="'.$this->createDateTime.'"
-            messageCount="1">
-            <EMandateStatusRequest entranceCode="'.$this->entranceCode.'">
-                <MandateID>'.$this->mandateID.'</MandateID>
-            </EMandateStatusRequest>
-        </EMandateInterface>';
-        */
-    }
-}
 
 /**
  * TransactionRequest
@@ -89,7 +27,7 @@ class EmandateBluemRequest extends BluemRequest
     protected $merchantID;
     protected $merchantSubID;
 
-    public function __construct($config, $customer_id, $order_id, $mandateID, String $expected_return="none")
+    public function __construct($config, $customer_id, $order_id, $mandateID, string $expected_return = "none")
     {
         parent::__construct($config, "", $expected_return);
 
@@ -112,7 +50,7 @@ class EmandateBluemRequest extends BluemRequest
         $this->mandateID = $mandateID;
 
         // https uniek returnurl voor klant
-        $this->merchantReturnURL = $this->merchantReturnURLBase."?mandateID={$this->mandateID}";
+        $this->merchantReturnURL = $this->merchantReturnURLBase . "?mandateID={$this->mandateID}";
         if (isset($config->sequenceType)) {
             $this->sequenceType = $config->sequenceType;
         } else {
@@ -136,8 +74,8 @@ class EmandateBluemRequest extends BluemRequest
         Wij presenteren het niet op de checkout, omdat wij zien dat veel partijen
         echt niet weten wat ze er in moeten zetten. Wij adviseren dan altijd klantnummer.
         En dat doet dan ook veel partijen */
-        if (isset($config->purchaseIDPrefix) && $config->purchaseIDPrefix!=="") {
-            $purchaseIDPrefix = $config->purchaseIDPrefix."-";
+        if (isset($config->purchaseIDPrefix) && $config->purchaseIDPrefix !== "") {
+            $purchaseIDPrefix = $config->purchaseIDPrefix . "-";
         } else {
             $purchaseIDPrefix = "";
         }
@@ -168,27 +106,28 @@ class EmandateBluemRequest extends BluemRequest
 
         $this->context = new MandatesContext($config->localInstrumentCode);
     }
-    public function XmlString() : String
+
+    public function XmlString(): string
     {
         return $this->XmlRequestInterfaceWrap(
             $this->xmlInterfaceName,
             'TransactionRequest',
             $this->XmlRequestObjectWrap(
                 'EMandateTransactionRequest',
-                '<MandateID>'.$this->mandateID.'</MandateID>
-                <MerchantReturnURL automaticRedirect="'.$this->automatically_redirect.'">'.$this->merchantReturnURL.'</MerchantReturnURL>
-                <SequenceType>'.$this->sequenceType.'</SequenceType>
-                <EMandateReason>'.$this->eMandateReason.'</EMandateReason>
-                <DebtorReference>'.$this->debtorReference.'</DebtorReference>
-                <PurchaseID>'.$this->purchaseID.'</PurchaseID>'.
-                $this->XmlWrapDebtorWallet().
+                '<MandateID>' . $this->mandateID . '</MandateID>
+                <MerchantReturnURL automaticRedirect="' . $this->automatically_redirect . '">' . $this->merchantReturnURL . '</MerchantReturnURL>
+                <SequenceType>' . $this->sequenceType . '</SequenceType>
+                <EMandateReason>' . $this->eMandateReason . '</EMandateReason>
+                <DebtorReference>' . $this->debtorReference . '</DebtorReference>
+                <PurchaseID>' . $this->purchaseID . '</PurchaseID>' .
+                $this->XmlWrapDebtorWallet() .
                 $this->XmlWrapDebtorAdditionalData(),
                 [
                     // 'entranceCode'=>$this->entranceCode,  always sent already
-                    'requestType'=>"Issuing",
-                    'localInstrumentCode'=>$this->localInstrumentCode,
-                    'merchantID'=>$this->merchantID,
-                    'merchantSubID'=>$this->merchantSubID,
+                    'requestType'         => "Issuing",
+                    'localInstrumentCode' => $this->localInstrumentCode,
+                    'merchantID'          => $this->merchantID,
+                    'merchantSubID'       => $this->merchantSubID,
                 ]
             )
         );
@@ -216,7 +155,7 @@ class EmandateBluemRequest extends BluemRequest
                 */
     }
 
-    public function TransactionType() : String
+    public function TransactionType(): string
     {
         return "TRX";
     }
