+++ src/Validators/EIdentityValidator.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Bluem\BluemPHP\Validators;
+
+/**
+ * Class EIdentityValidator
+ * @package Bluem\BluemPHP\Validators
+ *
+ * TODO.
+ */
+class EIdentityValidator extends Validator
+{
+    //public function __construct()
+    //{
+    //    parent::__construct(
+    //        __DIR__ . '/../validation/EIdentity.xsd'
+    //    );
+    //}
+}
