+++ src/Validators/EMandateValidator.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Bluem\BluemPHP\Validators;
+
+/**
+ * Class EMandateValidator
+ * @package Bluem\BluemPHP\Validators
+ *
+ * TODO.
+ */
+class EMandateValidator extends Validator
+{
+    //public function __construct()
+    //{
+    //    parent::__construct(
+    //        __DIR__ . '/../validation/EMandate.xsd'
+    //    );
+    //}
+}
