+++ src/Validators/EPaymentValidator.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Bluem\BluemPHP\Validators;
+
+/**
+ * Class EPaymentValidator
+ * @package Bluem\BluemPHP\Validators
+ *
+ * TODO.
+ */
+class EPaymentValidator extends Validator
+{
+    //public function __construct()
+    //{
+    //    parent::__construct(
+    //        __DIR__ . '/../validation/EPayment.xsd'
+    //    );
+    //}
+}
