+++ src/Validators/IBANCheckValidator.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Bluem\BluemPHP\Validators;
+
+/**
+ * Class IBANCheckValidator
+ * @package Bluem\BluemPHP\Validators
+ *
+ * TODO.
+ */
+class IBANCheckValidator extends Validator
+{
+    //public function __construct()
+    //{
+    //    parent::__construct(
+    //        __DIR__ . '/../validation/IBANCheck.xsd'
+    //    );
+    //}
+}
