+++ src/Validators/Validator.php
@@ -1,7 +1,9 @@
 <?php
 // Reference: https://www.codementor.io/@sirolad/validating-xml-against-xsd-in-php-6f56rwcds
 
-namespace Bluem\BluemPHP;
+namespace Bluem\BluemPHP\Validators;
+
+use Bluem\BluemPHP\Contexts\BluemContext;
 
 class Validator
 {
@@ -11,26 +13,35 @@ class Validator
      * @var string
      */
     protected $feedSchema = "";
+
     /**
      * @var int
      */
     public $feedErrors = 0;
+
     /**
      * Formatted libxml Error details
      *
      * @var array
      */
     public $errorDetails;
+
+    /**
+     * @var \DOMDocument
+     */
+    private $handler;
+
     /**
      * Validation Class constructor Instantiating DOMDocument
      *
-     * @param \DOMDocument $handler [description]
+     * @param string|null $feedSchema
      */
-    public function __construct(String $feedSchema = null)
+    public function __construct(string $feedSchema = null)
     {
         $this->handler = new \DOMDocument('1.0', 'utf-8');
         $this->feedSchema = $feedSchema;
     }
+
     /**
      * @param \libXMLError object $error
      *
@@ -40,30 +51,33 @@ class Validator
     {
         $errorString = "Error $error->code in $error->file (Line:{$error->line}):";
         $errorString .= trim($error->message);
+
         return $errorString;
     }
+
     /**
      * @return array
      */
     private function libxmlDisplayErrors()
     {
         $errors = libxml_get_errors();
-        $result    = [];
+        $result = [];
         foreach ($errors as $error) {
             $result[] = $this->libxmlDisplayError($error);
         }
         libxml_clear_errors();
+
         return $result;
     }
 
     /**
      * Validate Incoming Feeds against Listing Schema
      *
-     * @param resource $feeds
+     * @param BluemContext $context
+     * @param              $contents
      *
      * @return bool
-     *
-     * @throws \Exception
+     * @throws \DOMException
      */
     public function validate(BluemContext $context, $contents)
     {
@@ -71,12 +85,15 @@ class Validator
 
         if (!class_exists('DOMDocument')) {
             throw new \DOMException("'DOMDocument' class not found!");
+
             return false;
         }
         if (!file_exists($this->feedSchema)) {
             throw new \Exception('Schema is Missing, Please add schema to feedSchema property');
+
             return false;
         }
+
         libxml_use_internal_errors(true);
         // if (!($fp = fopen($feeds, "r"))) {
         //    die("could not open XML input");
@@ -87,7 +104,8 @@ class Validator
         $this->handler->loadXML($contents, LIBXML_NOBLANKS);
         if (!$this->handler->schemaValidate($this->feedSchema)) {
             $this->errorDetails = $this->libxmlDisplayErrors();
-            $this->feedErrors   = 1;
+            $this->feedErrors = 1;
+
             return false;
         } else {
             //The file is valid
@@ -95,6 +113,7 @@ class Validator
         }
 
     }
+
     /**
      * Display Error if Resource is not validated
      *
@@ -105,52 +124,3 @@ class Validator
         return $this->errorDetails;
     }
 }
-
-// /**
-//  * EIdentityValidator class
-//  */
-// class EIdentityValidator extends Validator
-// {
-//     public function __construct() {
-//         parent::__construct(
-//             __DIR__  .  '/../validation/EIdentity.xsd'
-//         );
-//     }
-// }
-
-
-// /**
-//  * EMandateValidator class
-//  */
-// class EMandateValidator extends Validator
-// {
-//     public function __construct() {
-//         parent::__construct(
-//             __DIR__  .  '/../validation/EMandate.xsd'
-//         );
-//     }
-// }
-
-// /**
-//  * EPaymentValidator class
-//  */
-// class EPaymentValidator extends Validator
-// {
-//     public function __construct() {
-//         parent::__construct(
-//             __DIR__  .  '/../validation/EPayment.xsd'
-//         );
-//     }
-// }
-
-// /**
-//  * IBANCheckValidator class
-//  */
-// class IBANCheckValidator extends Validator
-// {
-//     public function __construct() {
-//         parent::__construct(
-//             __DIR__  .  '/../validation/IBANCheck.xsd'
-//         );
-//     }
-// }
